Refactor 5.2 ccc

+++ wsdr/src/dsp/decimator.h
@@ -9,7 +9,7 @@ typedef struct rf_decimator rf_decimator;
 
 rf_decimator *decimator_init(rf_decimator_callback callback);
 
-int rf_decimator_set_parameters(rf_decimator *d, double sample_rate, int down_factor);
+int rf_decimator_set_parameters(rf_decimator *d, int sample_rate, int down_factor);
 
 int rf_decimator_decimate(rf_decimator *d, const cmplx_s32 *complex_signal, int len);
